Too many problems to continue. There are workarounds, so storing for future reference.
Read up a bit more... this convinced me to back off and choose a
different path:
http://crunchtools.com/comparison-linux-container-images/

We got similar problems. The dependency on `musl` makes it really hard
to reuse packages that are pre-built, which could be an option if this
were a real company and striving for the smallest ever, but does not
seem the right path with a yearly development cycle.